LayoutTests/fast/mediastream/RTCPeerConnection-icecandidate-event-expected.txt

Test RTCPeerConnection 'icecandidate' event and gathering done

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Local description set
PASS End of test promise chain
PASS Got candidate
PASS candidate instanceof RTCIceCandidate is true
PASS candidate.sdpMLineIndex is 0
FAIL candidate.sdpMid should be null (of type object). Was 1 (of type string).
PASS candidate.candidate: candidate:2013266431 1 udp 2013266432 192.168.0.100 38838 typ host generation 0
PASS Got candidate
PASS candidate instanceof RTCIceCandidate is true
PASS candidate.sdpMLineIndex is 0
FAIL candidate.sdpMid should be null (of type object). Was 1 (of type string).
PASS candidate.candidate: candidate:1019216383 1 tcp 1019216384 192.168.0.100 9 typ host tcptype passive generation 0
PASS Got candidate
PASS candidate instanceof RTCIceCandidate is true
PASS candidate.sdpMLineIndex is 0
FAIL candidate.sdpMid should be null (of type object). Was 1 (of type string).
PASS candidate.candidate: candidate:1677722111 1 tcp 1677722112 172.18.0.1 47989 typ srflx raddr 192.168.0.100 rport 47989 generation 0
PASS Got end of candidates
PASS candidate is null
PASS successfullyParsed is true

TEST COMPLETE
